feat: [Variables/Frontend] moved variables editing into VariablesEdit
